backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest==7.4.3